requirements.txt
aenum==2.1.2
aiocontextvars==0.2.1
aiofiles==0.4.0
aiosqlite==0.8.1
aniso8601==3.0.2
apispec==1.0.0rc1
apistar==0.6.0
asgiref==2.3.2
async-timeout==3.0.1
asyncpg==0.18.3
certifi==2018.11.29
chardet==3.0.4
ciso8601==2.1.1
Click==7.0
contextvars==2.3
docopt==0.6.2
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
h11==0.8.1
httptools==0.0.11
idna==2.8
immutables==0.6
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==2.18.0
parse==1.11.1
promise==2.2.1
PyPika==0.22.3
python-multipart==0.0.5
PyYAML==3.13
requests==2.21.0
requests-toolbelt==0.9.1
responder==1.2.0
rfc3986==1.2.0
Rx==1.6.1
six==1.12.0
starlette==0.10.0
tortoise-orm==0.11.2
urllib3==1.24.1
uvicorn==0.4.1
uvloop==0.12.0
websockets==7.0
whitenoise==4.1.2
